Cleanup of pull request, factored out ScalarExtensions. General cleanup
