+++ Peer.GitHub/GraphQL/PullRequestSearch/CheckRun.cs
@@ -50,5 +50,4 @@ namespace Peer.GitHub.GraphQL.PullRequestSearch
             };
         }
     }
-
 }
