--- Peer.GitHub/GraphQL/PullRequestSearch/Commits.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Peer.GitHub.GraphQL.PullRequestSearch
-{
-#nullable disable
-    public class Commits
-    {
-        public List<CommitNode> Nodes { get; set; }
-    }
-#nullable enable
-}
